ref(env-var): moved pkg from env_var to variables [#20]
